+++ samples/NoaaClientSample/Program.cs
@@ -1,7 +1,7 @@
-﻿using AuroraScienceHub.Integrations.NoaaClient;
+using AuroraScienceHub.Integrations.NoaaClient;
 using AuroraScienceHub.Integrations.NoaaClient.Ace;
-using AuroraScienceHub.Integrations.NoaaClient.Dscovr;
 using AuroraScienceHub.Integrations.NoaaClient.KpIndex;
+using AuroraScienceHub.Integrations.NoaaClient.Rtsw;
 using AuroraScienceHub.Integrations.Samples.NoaaClientSample;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -16,8 +16,8 @@ var host = builder.Build();
 
 // Get NOAA clients from DI
 var aceClient = host.Services.GetRequiredService<IAceClient>();
-var dscovrClient = host.Services.GetRequiredService<IDscovrClient>();
 var kpIndexClient = host.Services.GetRequiredService<IKpIndexClient>();
+var rtswClient = host.Services.GetRequiredService<IRtswClient>();
 
 // Display header
 AnsiConsole.Write(new FigletText("NOAA Client").Color(Color.Blue));
@@ -33,8 +33,8 @@ while (true)
             .AddChoices(
                 "ACE Magnetometer",
                 "ACE SWEPAM",
-                "DSCOVR Magnetometer",
-                "DSCOVR Solar Wind Plasma",
+                "RTSW Magnetometer",
+                "RTSW Solar Wind Plasma",
                 "KP Index 27-Day Forecast",
                 "KP Index 3-Day Forecast",
                 "KP Index Nowcast",
@@ -66,15 +66,15 @@ while (true)
                         var data = await aceClient.GetSwepamDataAsync(CancellationToken.None);
                         OutputFormatter.DisplayAceSwepam(data, 10);
                     }),
-                    "DSCOVR Magnetometer" => FetchAndDisplay(async () =>
+                    "RTSW Magnetometer" => FetchAndDisplay(async () =>
                     {
-                        var data = await dscovrClient.GetMagnetometerData2HAsync(CancellationToken.None);
-                        OutputFormatter.DisplayDscovrMagnetometer(data, 10);
+                        var data = await rtswClient.GetMagnetometerDataAsync(CancellationToken.None);
+                        OutputFormatter.DisplayRtswMagnetometer(data, 10);
                     }),
-                    "DSCOVR Solar Wind Plasma" => FetchAndDisplay(async () =>
+                    "RTSW Solar Wind Plasma" => FetchAndDisplay(async () =>
                     {
-                        var data = await dscovrClient.GetSolarWindPlasmaData2HAsync(CancellationToken.None);
-                        OutputFormatter.DisplayDscovrPlasma(data, 10);
+                        var data = await rtswClient.GetSolarWindPlasmaDataAsync(CancellationToken.None);
+                        OutputFormatter.DisplayRtswPlasma(data, 10);
                     }),
                     "KP Index 27-Day Forecast" => FetchAndDisplay(async () =>
                     {
@@ -132,15 +132,15 @@ async Task ExecuteAllRequests()
             var data = await aceClient.GetSwepamDataAsync(CancellationToken.None);
             OutputFormatter.DisplayAceSwepam(data, 5);
         }),
-        ("DSCOVR Magnetometer", async () =>
+        ("RTSW Magnetometer", async () =>
         {
-            var data = await dscovrClient.GetMagnetometerData2HAsync(CancellationToken.None);
-            OutputFormatter.DisplayDscovrMagnetometer(data, 5);
+            var data = await rtswClient.GetMagnetometerDataAsync(CancellationToken.None);
+            OutputFormatter.DisplayRtswMagnetometer(data, 5);
         }),
-        ("DSCOVR Solar Wind", async () =>
+        ("RTSW Solar Wind", async () =>
         {
-            var data = await dscovrClient.GetSolarWindPlasmaData2HAsync(CancellationToken.None);
-            OutputFormatter.DisplayDscovrPlasma(data, 5);
+            var data = await rtswClient.GetSolarWindPlasmaDataAsync(CancellationToken.None);
+            OutputFormatter.DisplayRtswPlasma(data, 5);
         }),
         ("KP Index Nowcast", async () =>
         {